requirements.txt
Flask
requests
python-dotenv
regex
Flask-Cors